VRC-4 - refactoring code, package
