+++ app/src/main/java/com/example/vrach/network/customer/CustomerApiService.kt
@@ -1,6 +1,5 @@
-package com.example.vrach.api
+package com.example.vrach.network.customer
 
-import com.example.vrach.model.Customer
 import okhttp3.RequestBody
 import okhttp3.ResponseBody
 import retrofit2.Response
@@ -8,7 +7,7 @@ import retrofit2.http.Body
 import retrofit2.http.POST
 
 
-interface CustomerApi {
+interface CustomerApiService {
     @POST("/v1/register")
     suspend fun createAccount(@Body requestBody: RequestBody): Response<ResponseBody>
 }
