+++ app/src/main/java/com/example/vrach/network/customer/CustomerInstance.kt
@@ -1,10 +1,10 @@
-package com.example.vrach.api
+package com.example.vrach.network.customer
 
 import com.example.vrach.utils.Constants.Companion.BASE_URL
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-object RetrofitInstance {
+object CustomerInstance {
     private val retrofit by lazy {
         Retrofit.Builder()
             .baseUrl(BASE_URL)
@@ -12,7 +12,7 @@ object RetrofitInstance {
             .build()
     }
 
-    val api: CustomerApi by lazy {
-        retrofit.create(CustomerApi::class.java)
+    val api: CustomerApiService by lazy {
+        retrofit.create(CustomerApiService::class.java)
     }
 }
